Mk2 is not required.
